Add visibility and birthdate fields, refactor playlist removal
Introduces IsVisible property to multiple domain entities for user visibility control. Adds Gender and BirthDate fields to artist and listener update requests and validators, with corresponding service logic. Refactors playlist track removal to use a new RemoveFromPlaylistRequest model and validator. Updates service interfaces and GraphQL queries to standardize method names. Adds manual user deletion mutation and service logic, and prevents deletion of 'Favorite Songs' playlist.

+++ EkofyApp.Infrastructure/Services/Playlists/PlaylistService.cs
@@ -193,12 +193,12 @@ public sealed class PlaylistService(IUnitOfWork unitOfWork, IHttpContextAccessor
         UpdateResult updateResult = await _unitOfWork.GetCollection<Playlist>().UpdateOneAsync(x => x.Id == favoritePlaylist.Id, updateDefinition);
     }
 
-    public async Task RemoveFromPlaylistAsync(AddToPlaylistRequest addToPlaylistRequest)
+    public async Task RemoveFromPlaylistAsync(RemoveFromPlaylistRequest removeFromPlaylistRequest)
     {
         UpdateDefinition<Playlist> updateDefinition = Builders<Playlist>.Update
-            .PullFilter(x => x.TracksInfo, y => y.TrackId == addToPlaylistRequest.TrackId);
+            .PullFilter(x => x.TracksInfo, y => y.TrackId == removeFromPlaylistRequest.TrackId);
         UpdateResult updateResult = await _unitOfWork.GetCollection<Playlist>()
-            .UpdateOneAsync(x => x.Id == addToPlaylistRequest.PlaylistId, updateDefinition);
+            .UpdateOneAsync(x => x.Id == removeFromPlaylistRequest.PlaylistId, updateDefinition);
 
         if (updateResult.ModifiedCount == 0)
         {
@@ -208,6 +208,14 @@ public sealed class PlaylistService(IUnitOfWork unitOfWork, IHttpContextAccessor
 
     public async Task DeletePlaylistAsync(string playlistId)
     {
+        if(await _unitOfWork.GetCollection<Playlist>()
+            .Find(x => x.Id == playlistId)
+            .Project(x => x.Name)
+            .FirstOrDefaultAsync() == "Favorite Songs")
+        {
+            throw new BadRequestCustomException("Cannot delete favorite songs playlist.");
+        }
+
         DeleteResult deleteResult = await _unitOfWork.GetCollection<Playlist>()
             .DeleteOneAsync(x => x.Id == playlistId);
